Restored sending of ttl as integer. JIRA CXX-7582
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70388 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/neticache_client.cpp
@@ -1091,7 +1091,7 @@ void CNetICacheClientExt::ProlongBlobLifetime(const string& key,
     cmd += "\" \"";
     cmd += key;
     cmd += "\" \"\" ttl=";
-    cmd += NStr::NumericToString(ttl.GetAsDouble());
+    cmd += NStr::NumericToString((unsigned)ttl.GetAsDouble());
 
     m_Impl->AppendClientIPSessionID(&cmd);
     m_Impl->ChooseServerAndExec(cmd, key, false, &m_Impl->m_DefaultParameters);
